+++ app/src/main/java/io/zoemeow/dutschedule/model/settings/BackgroundImageOption.kt
@@ -2,6 +2,6 @@ package io.zoemeow.dutschedule.model.settings
 
 enum class BackgroundImageOption(val value: Int) {
     None(0),
-    YourWallpaper(1),
-    ChooseFromFile(2)
+    YourCurrentWallpaper(1),
+    PickFileFromMedia(2)
 }